pyproject.toml
[project]
name = "dut-lowpowernotifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.51.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
